6.38
RECOVER MODE -> SQLITE.

+++ src/main/java/com/tonikelope/coronapoker/Stats.java
@@ -582,7 +582,7 @@ public class Stats extends javax.swing.JDialog {
 
             ResultSet rs = statement.executeQuery();
 
-            game_playtime_val.setText((rs.getObject("end") != null ? Helpers.seconds2FullTime((rs.getLong("end") / 1000 - rs.getLong("start") / 1000)) : "--:--:--") + " (" + Helpers.seconds2FullTime(rs.getInt("play_time")) + ")");
+            game_playtime_val.setText((rs.getObject("end") != null ? Helpers.seconds2FullTime((rs.getLong("end") / 1000 - rs.getLong("start") / 1000)) : "--:--:--") + " (" + Helpers.seconds2FullTime(rs.getLong("play_time")) + ")");
 
             String[] jugadores = rs.getString("players").split("#");
